+++ packages/backend/db/migrations/2022.11.20T11.57.37.sort_by_default.sql
@@ -1,0 +1,6 @@
+ALTER TABLE "collections" 
+ALTER COLUMN "sort_by" SET DEFAULT 'newestFirst';
+
+UPDATE "collections"
+SET "sort_by" = 'newestFirst'
+WHERE "sort_by" IS NULL;