+++ packages/frontend/api/collections.ts
@@ -7,6 +7,7 @@ import {
   ID,
   UpdateCollection,
   CollectionGrouping,
+  CollectionSortBy,
 } from '@orpington-news/shared';
 import type { Wretch } from '@api';
 
@@ -19,12 +20,13 @@ export const getCollectionItems = (
   collectionId: string | ID,
   pageIndex: number,
   filter: CollectionFilter,
-  grouping: CollectionGrouping
+  grouping: CollectionGrouping,
+  sortBy: CollectionSortBy
 ) =>
   api
     .options({ signal })
     .url(`/collections/${collectionId}/items`)
-    .query({ pageIndex, filter, grouping })
+    .query({ pageIndex, filter, grouping, sortBy })
     .get()
     .json<CollectionItem[]>();